create heartbeat led module
The heartbeat module is a somewhat generic timer-based heartbeat led.
Configuration is done through preprocesser #defines. This is because
there is not a good way to decouple the ISR from the module. The ISR
needs to know which timer and gpio port are being used. This information
can't easily be passed into the ISR without the heartbeat being a
singleton. One could do some poor man's C++ tricks (e.g., something like
the container_of macro in the Linux kernel) and make the timer isr call
a heartbeat callback function. Or one could define the ISR completely
outside of the module and make a different isr function for each
heartbeat led in the system. Neither of those options seemed very good
or worth the pain. A heartbeat LED is indeed generally a singleton as
most systems don't have more than one heartbeat led. A generic timer HAL
would be a better way to configure multiple blinky LEDs. This heartbeat
module is just a wrapper around a specific timer use case...